requirements.txt
python-dotenv
faiss-cpu 
sentence-transformers 
PyPDF2 
python-docx 
python-pptx 
streamlit 
ollama 
numpy
streamlit-lottie
requests